Support cookie-based RPC authentication (Bitcoin Core)
A constant `RPC_CREDENTIALS` is introduced that determines the RPC
authentication method: if credentials are set (i.e. 'Some(user, pass)'),
user-pass-authentication is used, otherwise (i.e. 'None'),
authentication is cookie-based (available since Bitcoin Core v0.12).

Of course, in the future the constant will change to a variable that is
assigned by the return value of a function that reads a configuration
file.

Note that by now the cookie-file is assumed to reside in
~/.bitcoin/regtest/.cookie and thus this authentication method will only
work for Linux and regtest network by now.