Make Wallet struct fields not public again
The fields in the wallet struct were made public in an earlier
commit so that they could be accessed in tests. Reverse this,
and add getter functions.

It is important that some of these fields are not accessed outside
the struct. Especially the swap_coins HashMap, which should be
backed up to disk whenever its added to. As an example see the
add_swapcoin function.