Add wake-up channel for server allowing shutdown
This commit has the server_socket.accept() be in a select!
statement along with a mpsc channel. By signalling the channel
a client loop can stop the server.

Right now this is only useful for simplying the integration test
as it allows ending the server in a different way. Later it will
be useful for error handling for errors that are fatal and require
the whole server to shut down.