Add some error handling to taker_protocol.rs
All the functions now use Result. In almost all cases the errors
are still just used with unwrap() in the main taker function, so
any unexpected result will just cause a crash.

Later the taker will be coded to change its behaviour when an error
happens. For example if a maker doesnt reply then the taker will
try to find another one which does.