Implement first version of watchtower
This watchtower for now will broadcast all the contract transactions
if any of them are seen on the network, either in the mempool or in
blocks. The security model of multi-transaction coinswap requires
that if even one contract transaction is broadcasted then all of them
must be. Note this watchtower is only intended to be used by makers.
Takers will watch the network during a coinswap and have to respond
in a different way, by stopping a multi-hop coinswap for example.

Also implemented a `test-watchtower-client` new subcommand. This allows
trying out the watchtower without starting up a maker and doing a whole
coinswap.

Later the watchtower will also need to read a hashvalue from transactions
and use that hashvalue to create their own transactions. This isnt
implemented yet.