Have sign_transaction() use own enum for signing
Previously the sign_transaction function would require a psbt json Value
previously obtained from the wallet_create_funded_psbt RPC call. That meant
the function couldnt be used without the psbt. A better way is for the
function to accept an iterator of enums which give the required information
for signing a transaction.