Have taker repeatedly attempt to connect to makers
Previously the taker code would simply give up after the first failure. So even
a slightly choppy internet connection would cause a failure.

This commit only does this for connections that request the senders contract tx signatures,
that request the receivers contract tx sigs, and connections which do the hash preimage
revealing and private key handover. The only thing remaining is the connection which sends
the ProofOfFunding message, but thats more complicated and will be done in another commit.

This robustness is important for decentralization, without it only makers on very
stable servers and connections which never go down would earn the max profit.
We want makers to be runnable on raspberry pis in people's bedrooms, that would be
out without more robustness and re-trying the connections.