Add alternative function for creating funding txes
Based on my dogfooding which showed that sometimes the code is unable
to properly choose UTXOs to create funding transactions. So a fallback
is needed and maybe more than one.

This commit renames the functions create_spending_txes -> create_funding_txes

This commit also moves many of those routines to a new file funding_tx.rs
as wallet_sync.rs was getting pretty big, and going forward the creation
of funding transactions will have to be very careful in order to get
the best privacy.