Check RPC connection before rejecting all clients
Previously if an RPC error arose in maker or watchtower then it
would interpret that as if the RPC connection had failed, and reject
all new incoming clients until it managed to re-establish.

However sometimes an RPC error would arise because of other reasons
than connection failure, for example "insufficent funds", and then it
makes no sense to reject all new clients.

Because of this, it makes sense to check again the RPC connection
before rejecting all clients.