Check RPC ping interval seperately from heartbeat
Previously the maker heartbeat would be exactly the RPC ping interval.
As the RPC ping interval was fairly long (60 seconds), if a client connected
sooner, then the heartbeat would be reset back to the start because of how
the select loop works. If a client repeatedly reconnect more frequentyly than
60 seconds it would result in the heartbeat code never being reached.