Taker Refactoring.
Create a dedicated Taker struct, which includes a OngoingSwapState. This
state is used to perform all the swap protocol specific operations. These
states are specific to a single swap and are ephemeral. Before the finalization
of the swap the state is cashed into wallet file.

Various other utility structures are added.

Taker configuration structure added which defaults to the hard code variables.